Move notifyDataImportProgress
