+++ src/pages/ImportData.js
@@ -3,12 +3,12 @@ import { Fragment, useCallback, useState } from 'react';
 import { div, h, h2, img, li, p, span, strong, ul } from 'react-hyperscript-helpers';
 import Collapse from 'src/components/Collapse';
 import { ButtonPrimary, ButtonSecondary, Clickable, IdContainer, Link, RadioButton, spinnerOverlay } from 'src/components/common';
-import { notifyDataImportProgress } from 'src/components/data/data-utils';
 import FooterWrapper from 'src/components/FooterWrapper';
 import { icon, wdlIcon } from 'src/components/icons';
 import NewWorkspaceModal from 'src/components/NewWorkspaceModal';
 import TopBar from 'src/components/TopBar';
 import { useWorkspaces, WorkspaceSelector } from 'src/components/workspace-utils';
+import { notifyDataImportProgress } from 'src/data/import-jobs';
 import jupyterLogo from 'src/images/jupyter-logo.svg';
 import scienceBackground from 'src/images/science-background.jpg';
 import { Ajax } from 'src/libs/ajax';
